Test refactor:FragmentRunner and ProtocolFragment moved to their own file for use in tests
