+++ test/show_protocol.py
@@ -4,22 +4,20 @@ import sys
 import argparse
 import time
 import typing
-from typing import Optional, List, Sequence, Dict
+from typing import Optional, List, Sequence
 
 import dotenv
 
 import load_house
 from actors.actor_base import ActorBase
 from actors.atn import Atn
-from actors.boolean_actuator import BooleanActuator
 from actors.cloud_ear import CloudEar
-from actors.power_meter import PowerMeter
-from actors.simple_sensor import SimpleSensor
 from command_line_utils import add_default_args, setup_logging
 from config import ScadaSettings
 from data_classes.sh_node import ShNode
 from drivers.power_meter.gridworks_sim_pm1__power_meter_driver import GridworksSimPm1_PowerMeterDriver
-from test.utils import ScadaRecorder, AtnRecorder, HomeAloneRecorder, wait_for
+from fragment_runner import ProtocolFragment, FragmentRunner, do_nothing, delimit
+from utils import wait_for
 
 # noinspection PyUnusedLocal
 def i_am_quiet(self, note: str):
@@ -38,6 +36,8 @@ class FragmentNames(enum.Enum):
     meter = "meter"
     GsPwr = "GsPwr"
     thermo = "thermo"
+    relay = "relay"
+    relay_toggle = "relay_toggle"
 
 def add_show_protocol_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
     """Add show_protocol only args"""
@@ -52,143 +52,51 @@ def add_show_protocol_args(parser: argparse.ArgumentParser) -> argparse.Argument
     parser.add_argument("-w", "--wait-at-least", default=0, help="Time to wait for all periodic fragments", type=float)
     return parser
 
-def delimit_str(text: str = "") -> str:
-    return "\n## " + text + ("#" * (100 - len(text)))
-
-def delimit(text: str = ""):
-    print(delimit_str(text))
-
-def do_nothing(seconds: float):
-    """Let the actors run on their own for a while"""
-    if seconds > 0:
-        delimit(f"DOING NOTHING FOR {int(seconds):4d} SECONDS")
-        time.sleep(seconds)
-        delimit("DONE DOING NOTHING")
-
-class Actors:
-    scada: ScadaRecorder
-    atn: AtnRecorder
-    home_alone: HomeAloneRecorder
-    relay: BooleanActuator
-    meter: PowerMeter
-    thermo: SimpleSensor
-
-    def __init__(self, settings:ScadaSettings):
-        self.scada = ScadaRecorder(node=ShNode.by_alias["a.s"], settings=settings)
-        self.atn = AtnRecorder(node=ShNode.by_alias["a"], settings=settings)
-        self.home_alone = HomeAloneRecorder(node=ShNode.by_alias["a.home"], settings=settings)
-        self.relay = BooleanActuator(ShNode.by_alias["a.elt1.relay"], settings=settings)
-        self.meter = PowerMeter(node=ShNode.by_alias["a.m"], settings=settings)
-        self.thermo = SimpleSensor(node=ShNode.by_alias["a.tank.temp0"], settings=settings)
-
-
-class FragmentRunner:
-    args: argparse.Namespace
-    settings: ScadaSettings
-    actors: Actors
-    requested: Dict[str, ActorBase]
-    fragments: List["ProtocolFragment"]
-    wait_at_least: float
-
-    def __init__(self, args:argparse.Namespace, settings:ScadaSettings, actors:Optional[Actors] = None):
-        self.args = args
-        self.settings = settings
-        self.actors = Actors(settings) if actors is None else actors
-        self.requested = dict()
-        self.fragments = []
-        self.wait_at_least = args.wait_at_least
-        for fragment in [self.fragment_from_enum(FragmentNames(name)) for name in self.args.fragments]:
-            self.add_fragment(fragment)
-
-    def add_fragment(self, fragment: "ProtocolFragment") -> "FragmentRunner":
-        self.fragments.append(fragment)
-        self.wait_at_least = max(self.wait_at_least, fragment.wait_at_least)
-        self.request_actor(fragment.get_requested_actors())
-        return self
-
-    def fragment_from_enum(self, name: FragmentNames) -> "ProtocolFragment":
-        if name == FragmentNames.all:
-            fragment = AllFragments(self)
-        elif name == FragmentNames.thermo:
-            fragment = ThermoFragment(self)
-        elif name == FragmentNames.meter:
-            fragment = MeterFragment(self)
-        elif name == FragmentNames.GsPwr:
-            fragment = GsPwrFragment(self)
-        else:
-            raise ValueError(f"Unknown fragment name {name}")
-        return fragment
-
-    def request_actor(self, actors: Sequence[ActorBase]) -> "FragmentRunner":
-        for actor in actors:
-            if actor.node.alias not in self.requested:
-                self.requested[actor.node.alias] = actor
-        return self
-
-    def start(self):
-        for actor in self.requested.values():
-            actor.start()
-
-    def wait_connect(self):
-        for actor in self.requested.values():
-            if hasattr(actor, "client"):
-                wait_for(
-                    actor.client.is_connected,
-                    1,
-                    tag=f"ERROR waiting for {actor.node.alias} client connect",
-                )
-            if hasattr(actor, "gw_client"):
-                wait_for(
-                    actor.gw_client.is_connected,
-                    1,
-                    "ERROR waiting for gw_client connect",
-                )
-
-    def stop(self):
-        for actor in self.requested.values():
-            # noinspection PyBroadException
-            try:
-                actor.stop()
-            except:
-                pass
-
-    def run(self, *args, **kwargs):
-        try:
-            start_time = time.time()
-            delimit("STARTING")
-            self.start()
-            self.wait_connect()
-            delimit("CONNECTED")
-            for fragment in self.fragments:
-                fragment.run(*args, **kwargs)
-            if (time_left := self.wait_at_least - (time.time() - start_time)) > 0:
-                do_nothing(time_left)
-        finally:
-            self.stop()
-
-class ProtocolFragment:
-    runner: FragmentRunner
-    wait_at_least: float
-
-    def __init__(self, runner:FragmentRunner, wait_at_least: float = 0):
-        self.runner = runner
-        self.wait_at_least = wait_at_least
-
-    def get_requested_actors(self) -> Sequence[ActorBase]:
-        pass
-
-    def run(self, *args, **kwargs):
-        pass
+def fragment_from_enum(runner: FragmentRunner, name: FragmentNames) -> "ProtocolFragment":
+    if name == FragmentNames.all:
+        fragment = AllFragments(runner)
+    elif name == FragmentNames.thermo:
+        fragment = ThermoFragment(runner)
+    elif name == FragmentNames.meter:
+        fragment = MeterFragment(runner)
+    elif name == FragmentNames.GsPwr:
+        fragment = GsPwrFragment(runner)
+    elif name == FragmentNames.relay:
+        fragment = RelayFragment(runner)
+    elif name == FragmentNames.relay_toggle:
+        fragment = RelayToggleFragment(runner)
+    else:
+        raise ValueError(f"Unknown fragment name {name}")
+    return fragment
 
 class WaitingFragment(ProtocolFragment):
 
     def __init__(self, runner:FragmentRunner):
-        super().__init__(runner, wait_at_least = runner.args.wait_at_least)
+        super().__init__(runner, wait_at_least = runner.wait_at_least)
 
 class ThermoFragment(WaitingFragment):
 
     def get_requested_actors(self) -> Sequence[ActorBase]:
         return [self.runner.actors.scada, self.runner.actors.thermo]
+
+class RelayFragment(WaitingFragment):
+
+    def get_requested_actors(self) -> Sequence[ActorBase]:
+        return [self.runner.actors.scada, self.runner.actors.relay]
+
+class RelayToggleFragment(ProtocolFragment):
+
+    def get_requested_actors(self) -> Sequence[ActorBase]:
+        return [self.runner.actors.scada, self.runner.actors.relay]
+
+    def run(self, *args, **kwargs):
+        time.sleep(1)
+        scada = self.runner.actors.scada
+        relay = self.runner.actors.relay
+        scada.turn_on(relay.node)
+        time.sleep(1)
+        scada.turn_off(relay.node)
+        time.sleep(1)
 
 class MeterFragment(WaitingFragment):
 
@@ -227,17 +135,16 @@ class AllFragments(ProtocolFragment):
 
     def run(self, *args, **kwargs):
         actors = self.runner.actors
-        args = self.runner.args
         actors.scada._scada_atn_fast_dispatch_contract_is_alive_stub = True
 
-        do_nothing(args.do_nothing_time)
+        do_nothing(self.runner.do_nothing_time)
 
         delimit("TURNING ON")
         actors.atn.turn_on(ShNode.by_alias["a.elt1.relay"])
         wait_for(lambda: actors.relay.relay_state == 1, 10, f"Relay state")
         delimit("TURNED ON")
 
-        do_nothing(args.do_nothing_time)
+        do_nothing(self.runner.do_nothing_time)
 
         delimit("REQUESTING STATUS")
         actors.atn.status()
@@ -262,7 +169,7 @@ class AllFragments(ProtocolFragment):
         )
         delimit("SCADA GOT STATUS")
 
-        do_nothing(args.do_nothing_time)
+        do_nothing(self.runner.do_nothing_time)
 
         delimit("SCADA SENDING STATUS")
         actors.scada.send_status()
@@ -285,9 +192,15 @@ def show_protocol(argv: Optional[List[str]] = None):
     setup_logging(args, settings)
     please_be_quiet()
     load_house.load_all(settings.world_root_alias)
+    runner = FragmentRunner(
+        settings,
+        wait_at_least=args.wait_at_least,
+        do_nothing_time=args.do_nothing_time,
+    )
     if not args.fragments:
         args.fragments = ["all"]
-    runner = FragmentRunner(args, settings)
+    for fragment_name in args.fragments:
+        runner.add_fragment(fragment_from_enum(runner, FragmentNames(fragment_name)))
     runner.run()
 
 if __name__ == "__main__":
